.history/requirements_20231017200458.txt
flask
bs4
requests
PyPDF2
PyYAML
lxml
